server/requirements.txt
# Django核心框架
Django~=4.2.0
djangorestframework==3.14.0
drf-dynamic-fields==0.4.0
drf-yasg==1.21.7
django-filter~=23.2
django-cors-headers>=4.0.0
channels>=4.0.0  # WebSocket支持
channels-redis>=4.0.0  # Channels Redis后端

# 数据库支持
psycopg2-binary>=2.9.0  # PostgreSQL支持（可选）
# mysqlclient==2.2.0  # MySQL支持

# HTTP和网络请求
requests~=2.31.0
urllib3~=2.0.4
certifi~=2023.7.22

# 数据处理
PyYAML==6.0.1
six~=1.16.0
pandas>=2.0.0
numpy>=1.24.0

# 日志和调试
loguru~=0.7.2

# 服务器
uvicorn~=0.24.0
gunicorn>=21.0.0  # 生产环境WSGI服务器
daphne>=4.0.0  # ASGI服务器

# 工具库
setuptools~=68.0.0
python-dotenv>=1.0.0  # 环境变量管理

# AI和机器学习
openai>=1.0.0  # OpenAI API
langchain>=0.1.0  # LangChain框架
langchain-openai>=0.0.5  # LangChain OpenAI集成
langgraph>=0.0.20  # 多智能体工作流
chromadb>=0.4.0  # 向量数据库

# 市场数据源
akshare>=1.11.0  # A股数据
tushare>=1.3.0  # Tushare数据
yfinance>=0.2.0  # Yahoo Finance数据

# 缓存和消息队列
redis>=5.0.0  # Redis客户端
celery>=5.3.0  # 异步任务队列（可选）

# 其他依赖
Pillow>=10.0.0  # 图像处理
inflection>=0.5.1  # 命名转换
